pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["ninovrolijk <nino.vrolijk@incentro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.5"
pandas = "^2.2.2"
requests = "^2.32.3"
datetime = "^5.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
